Switch to using resume scss files
This removes the existing css files from the theme and changes to using
the scss files from my resume site. This is possibly a very bad idea,
but I'm going to work from here to match some of the markup so that I
can try and have some consistency.

In the process, I had to fix up the Dockerfile. Turns out the scss
building features of Hugo require 'hugo extended', and hugo extended in
turn doesn't like being built on Alpine because of dependencies.