Add breakpoints mixin, with use in header
Need to make this site responsive, so I added in my standard
breakpoints scss mixin. Adjusted the breakpoints slightly, I think the
laptop-l one was too big, so shifted it down to the chrome laptop
breakpoint.

Made use of the mixin in the header scss partial to reduce the margin
and padding a bit.