Factor out some partials from the single page layout
This factors out some sections of the single page layout into partials.
My original idea for laying out recipes was to create a recipe layout,
but instead I opted for shortcodes because I wanted the markdown of a
recipe to better reflect the final page structure, as opposed to having
to put all of the detail of the recipe in parameters at the top of the
new layout.

Regardless, there's no harm in having these bits as partials, so I'm
going to leave them there.