Don't show the final hr in the final recipe block
This allows us to have multiple recipe blocks in a post. The first
recipe will have an <hr> below it. This allows for further explanation
between recipes. The final recipe will not have an <hr> below it, if it
is the final element in the post (i.e.) no further explanation below it.
This should allow us to avoid a double hr at the bottom of recipes.