Misc cleanup of the README and docker files
In my old age I'm coming to hate Docker more and more for stuff like
this. Installing a binary is a lot easier, especially when Podman
doesn't want to cooperate.

This removes the dockerfiles (which really only made sense for the
complexity of building the theme, which is going away soon anyway) and
replaces info about them with just downloading the binary and installing
it.

I also added some TODOs to the readme to remind me to do stuff.