Update theme and re-render using latest Hugo
This updates the blog to be rendered by the latest version of Hugo
extended. I was running a version that was a few years old, this updates
one spot where we needed to change a Hugo directive in the theme, and
then re-renders the whole site using the new version.

I don't 100% understand why the diff is so large, but none of it is
particularly worrying and I've spot checked pages and they don't look
terribly different visually.