Finish of the theme redesign
This finishes off my redesign.
All of the old custom CSS and SASS has been removed in favor of
Tailwind, and everything is looking more or less how I want it to.

Notable changes:
* Tailwind has a great 'prose' utility which I'm using for the actual
  posts themselves, customized with a little green lefthand border
  thingo
* I managed to eventually figure out how to clip the images to the right
  size, which I'm pretty happy with
* I changed the recipe shortcode to take an optional parameter as to
  whether to render a lower horizontal rule, so that we can switch this
  off if a recipe is at the end of a post. Would be nice to make this
  automatic in the future.
* Images on the list page now have alt text, thanks to a new frontmatter
  field.

There are still some todos in the README, but I think this is good to
go!