Rejig how I scale images to the aspect ratio in a list view
Previously this was using an ugly method of getting the images to fit
the desired aspect ratio involving translating and setting the images to
be positioned half way through their container.

This method is much cleaner, and works in all situations no matter what
the size of the image is, which is great.