Add index and finish off redesign posts
This adds an index as a landing page. Since it was kind of relevant, I
finished off the skouf.com redesign posts and published them as well,
since they've been sitting in the can for so long they were starting to
get stale, and I thought it was worth publishing while they were fresh
in my head.