Add links to some of my external articles
This adds a few links to my favourite articles that I've written on the
SEEK blog. This is accomplished via a new bit of logic in the list item
partial that detects whether a post is 'external content' via some
frontmatter and links to the external content instead of the permalink
for the leaf page.

I also updated my about section a little to streamline it.