Update index page to have a link to all posts
This is s fairly small change in that the index page now links off to
the list of posts with a simple 'all posts' link at the bottom.
I also added a conditional to the layout for where that link goes so
that instead of the heading being 'Posts posts', it says 'All posts'.
This resulted in new CSS, hence the re-render of every page that loads
the cache-busted CSS.